Add EfCore mapping and dapper

+++ PetFamily.Backend/src/PetFamily.Application/Models/PagedList.cs
@@ -3,7 +3,7 @@ namespace PetFamily.Application.Models;
 public class PagedList<T>
 {
     public IReadOnlyList<T> Items { get; init; } = [];
-    public int TotalCount { get; init; }
+    public long TotalCount { get; init; }
     public int Page { get; init; }
     public int PageSize { get; init; }
     public bool HasNextPage => TotalCount > Page * PageSize;
